Add leaderboard and gamification flow
Introduce leaderboard endpoints and integrate gamification into attempt submission. Added leaderboardController and leaderboardRoutes, registered the route in index.js, and updated response helper to support pagination. Implemented Attempt.recalculateScoreWithGamification to recompute score inside a transaction and invoke gamificationService.process, which updates XP, determines rank changes, and awards badges. Updated attemptController to use the new method and include xp/level/badge info in responses (with 422 error handling). Adjusted tests to mock the new flow and added property-based tests for determineRank; also switched Fast-Check floats to doubles where needed.

+++ backend/express_logilearn/test/properties/calculateXp.property.test.js
@@ -14,7 +14,7 @@ describe('calculateXp — Property-Based Tests', () => {
   it('xp_gained selalu dalam [0, 100] untuk skor valid', () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 0, max: 100, noNaN: true }),
+        fc.double({ min: 0, max: 100, noNaN: true }),
         (score) => {
           const xp = calculateXp(score);
           return xp >= 0 && xp <= 100 && xp === Math.floor(score);
@@ -30,8 +30,8 @@ describe('calculateXp — Property-Based Tests', () => {
     fc.assert(
       fc.property(
         fc.oneof(
-          fc.float({ max: -0.001, noNaN: true }),
-          fc.float({ min: 100.001, max: 1e6, noNaN: true })
+          fc.double({ max: -0.001, noNaN: true }),
+          fc.double({ min: 100.001, max: 1e6, noNaN: true })
         ),
         (invalidScore) => {
           expect(() => calculateXp(invalidScore)).toThrow();
